refactor: split ui_app.cpp into focused translation units
1038-line monolith → 7 files via a new internal header ui_render.hpp:

  ui_render.hpp   — inline helpers (to_v4, to_u32, ws, local_tm),
                    debug macro + declaration, render_* declarations
  ui_titlebar.cpp — render_titlebar, debug_set_ui_scale
  ui_clock.cpp    — render_clock + static helpers
  ui_stopwatch.cpp— render_stopwatch
  ui_timers.cpp   — render_timers + static helpers
  ui_alarms.cpp   — render_alarms, render_add_alarm_window
  ui_settings.cpp — render_settings_window, open_settings
  ui_app.cpp      — render_app, apply_imgui_theme, debug infra (~150 lines)

Also fixes two latent bugs:
- g_debug_overlay_visible references in render_clock corrected to
  g_debug_flag && *g_debug_flag (compile error under CHRONOS_DEBUG_UI_OVERLAY=ON)
- Three raw #ifdef _WIN32 localtime_s/r blocks in ui_app.cpp replaced with
  local_tm() from the new internal header (same consolidation as main_loop Fix 7)

main_loop.cpp: drops its static local_tm(); includes ui_render.hpp instead.
