+++ src/platform/main_loop.cpp
@@ -6,19 +6,11 @@
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
-#include <ctime>
 #include <filesystem>
-#include <format>
 #include <string>
 #include <vector>
-#include "actions.hpp"
-#include "alarm.hpp"
-#include "app.hpp"
 #include "config_io.hpp"
-#include "encoding.hpp"
-#include "platform_window.hpp"
-#include "ui_style.hpp"
-#include "ui_app.hpp"
+#include "ui_render.hpp"
 
 // ─── Screenshot helper ────────────────────────────────────────────────────────
 
@@ -42,16 +34,6 @@ static bool save_screenshot(const char* path, int w, int h) {
     return stbi_write_png(path, w, h, 3, pixels.data(), w * 3) != 0;
 }
 #endif
-
-static tm local_time(time_t t) {
-    tm lt{};
-#ifdef _WIN32
-    localtime_s(&lt, &t);
-#else
-    localtime_r(&t, &lt);
-#endif
-    return lt;
-}
 
 static void render_frame(SDL_Window* window, SDL_GLContext gl_ctx, App& app, UiState& ui) {
     SDL_GL_MakeCurrent(window, gl_ctx);
@@ -96,7 +78,7 @@ static bool process_due_notifications(SDL_Window* window, App& app) {
         }
     }
 
-    tm lt = local_time(std::time(nullptr));
+    tm lt = local_tm(std::time(nullptr));
     int h = lt.tm_hour, m = lt.tm_min;
     int cur_min = h * 60 + m;
     if (cur_min != app.alarm_notified_minute) {
@@ -239,7 +221,7 @@ int main(int argc, char* argv[]) {
                 ui.alarm_hour = 8; ui.alarm_minute = 0;
                 ui.alarm_days_mode = true;
                 for (int d = 0; d < 7; ++d) ui.alarm_days[d] = true;
-                tm lt = local_time(std::time(nullptr));
+                tm lt = local_tm(std::time(nullptr));
                 ui.alarm_year  = lt.tm_year + 1900;
                 ui.alarm_month = lt.tm_mon + 1;
                 ui.alarm_day   = lt.tm_mday;
